added support for connection.sync flag
Added support for a widely believed feature of Strophe that doesn't seem to actually be supported.

Setting connection.sync = true is believed to force synchronous connections according
to many blog and forum posts.  The source doesn't seem to support this at all. This little
patch makes that vaporware feature a reality.